Avancement

+++ src/ChainePositionnee.java
@@ -1,5 +1,5 @@
 public class ChainePositionnee{
-    int x,y;
+    double x,y;
     String c;
-    public ChainePositionnee(int a,int b, String d){x=a; y=b; c=d;}
+    public ChainePositionnee(double a,double b, String d){x=a; y=b; c=d;}
 }
